Adds currency conversion method and date sorting

+++ src/main/java/ui/Ui.java
@@ -306,7 +306,7 @@ public class Ui {
 
     }
 
-    public static void subFromSavings(int amount, int currentAmount) {
+    public static void subFromSavings(double amount, double currentAmount) {
         showLine();
         System.out.println("Subtracted " + amount + " from your savings.");
         if (currentAmount < 0) {
@@ -315,7 +315,7 @@ public class Ui {
         showLine();
     }
 
-    public static boolean printGoalStatus(int currentAmount, int targetAmount) {
+    public static boolean printGoalStatus(double currentAmount, double targetAmount) {
         showLine();
         if (currentAmount >= targetAmount) {
             System.out.println("You have achieved the goal! Congratulations!");
